requirements.txt
requests>=2.31.0,<3.0
pandas>=2.1.0,<3.0
python-dotenv>=1.0.0,<2.0
fastapi>=0.111.0,<1.0
uvicorn[standard]>=0.29.0,<1.0
jinja2>=3.1.0,<4.0
# Optional: enables Azure Table Storage audit log when AZURE_STORAGE_CONNECTION_STRING is set
azure-data-tables>=12.5.0,<13.0
